Remove redundant file and refine file structures
Removed `main_avg.py` in the reinforcement learning bandit problem folder because it was redundant. Renamed `main.py` to `stationary_bandit_epsilon_greedy.py` and moved it to examples directory. Also, moved the agent classes (`EpsilonGreedyAgent` and `AlphaEpsilonGreedyAgent`) to separate files under a new `agents` directory for better organization.

Moreover, updated the dependencies in `pyproject.toml`, and the settings in `mypy.ini`, along with task configurations in `taskipy.tasks` section for better automatic checking. Other changes include some code refinements in the classes of `GridWorld` and `EpsilonGreedyAgentBase`, for enhancing the robustness.

These changes aim to make the project's file architecture clearer and more standard, making it easier for others to understand and contribute to the codebase. The change in dependencies supports the newer versions and enhances the task automation.