Update Python Docstrings and clean up code formatting
In this commit, the Docstrings of various Python scripts were updated to follow the Google Python style recommendation of including a colon after 'Returns' and 'Attributes' to provide more clear and well-structured documentation. Also, cleaning up formatting of some Python script lines and methods were done for better organization such as, reformatting the way the variables are assigned, rearranging the line breaks, and removing unnecessary parentheses. In addition to the Python scripts, linter rules and other configurations were added to the 'ruff.toml' file and 'poetry.lock' file was reformatted by an automated tool. Both changes are made for better adherence to best practices.