Improve test accuracy with '--no-cov-on-fail' command
In this change, the command for running the unittests within the CI workflow has been updated. Previously, the pytest command included '--cov' to generate a coverage report and 'term-missing:skip-covered' to exclude covered lines from the report. The previous command is modified with the addition of '--no-cov-on-fail' flag.
This changes improves the accuracy of the test reports as pytest won't fail immediately due to coverage issues and will continue executing the remaining tests. This allows us to identify all the failing tests in a single run rather than rerunning the tests multiple times.